# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Estate Agent CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Contact {
  id: number;
  name: string;
  email?: string | null;
  phone: string;
  tags: string[];
  /** buyer, renter, landlord, vendor */
  leadIntent?: string | null;
  /** Summary of what the prospect is looking for */
  summary?: string | null;
  propertyInMind?: string | null;
  bedrooms?: number | null;
  budget?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateContactRequest {
  name: string;
  email?: string | null;
  phone: string;
  tags?: string[];
  leadIntent?: string | null;
  summary?: string | null;
  propertyInMind?: string | null;
  bedrooms?: number | null;
  budget?: string | null;
}

export interface UpdateContactRequest {
  name?: string;
  email?: string | null;
  phone?: string;
  tags?: string[];
  leadIntent?: string | null;
  summary?: string | null;
  propertyInMind?: string | null;
  bedrooms?: number | null;
  budget?: string | null;
}

export interface Conversation {
  id: number;
  contactId: number;
  contactName: string;
  contactPhone: string;
  /** whatsapp, sms, email */
  channel: string;
  /** active, closed, pending */
  status: string;
  lastMessage?: string | null;
  lastMessageAt?: string | null;
  unreadCount: number;
  createdAt: string;
}

export interface Message {
  id: number;
  conversationId: number;
  /** inbound, outbound */
  direction: string;
  content: string;
  sentAt: string;
  senderName?: string | null;
}

export interface ConversationDetail {
  id: number;
  contactId: number;
  contactName: string;
  contactPhone: string;
  channel: string;
  status: string;
  lastMessage?: string | null;
  lastMessageAt?: string | null;
  unreadCount: number;
  createdAt: string;
  messages: Message[];
}

export interface CreateConversationRequest {
  contactId: number;
  channel: string;
}

export interface SendMessageRequest {
  content: string;
}

export interface DailyCount {
  date: string;
  count: number;
}

export interface ChannelCount {
  channel: string;
  count: number;
}

export interface FunnelData {
  newLeads: number;
  engaged: number;
  qualified: number;
  viewingsBooked: number;
  offers: number;
}

export interface DashboardMetrics {
  leadsEngaged: number;
  leadsEngagedInstantly: number;
  viewingsBooked: number;
  reactivatedLeads: number;
  referralLeads: number;
  portalLeadsQualified: number;
  hotLeads: number;
  buyers: number;
  renters: number;
  landlordsVendors: number;
  costSavingsGBP: number;
  timeSavingsHours: number;
  newConversationsLastWeek: DailyCount[];
  messagesByChannel: ChannelCount[];
  funnel: FunnelData;
}

export type ListContactsParams = {
  search?: string;
  tag?: string;
  leadIntent?: string;
};

export type ListConversationsParams = {
  search?: string;
  channel?: string;
};