# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Estate Agent CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: number;
  name: string;
  email?: string | null;
  phone: string;
  tags: string[];
  /** buyer, renter, landlord, vendor */
  leadIntent?: string | null;
  /** Summary of what the prospect is looking for */
  summary?: string | null;
  propertyInMind?: string | null;
  bedrooms?: number | null;
  budget?: string | null;
  createdAt: Date;
  updatedAt: Date;
}