# lib/api-zod/src/generated/types/conversationDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Estate Agent CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from "./message";

export interface ConversationDetail {
  id: number;
  contactId: number;
  contactName: string;
  contactPhone: string;
  channel: string;
  status: string;
  lastMessage?: string | null;
  lastMessageAt?: Date | null;
  unreadCount: number;
  createdAt: Date;
  messages: Message[];
}