# lib/api-zod/src/generated/types/createConversationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Estate Agent CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateConversationRequest {
  contactId: number;
  channel: string;
}