# lib/api-zod/src/generated/types/funnelData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Estate Agent CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface FunnelData {
  newLeads: number;
  engaged: number;
  qualified: number;
  viewingsBooked: number;
  offers: number;
}