# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Estate Agent CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  conversationId: number;
  /** inbound, outbound */
  direction: string;
  content: string;
  sentAt: Date;
  senderName?: string | null;
}