# lib/api-zod/src/generated/types/updateContactRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Estate Agent CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateContactRequest {
  name?: string;
  email?: string | null;
  phone?: string;
  tags?: string[];
  leadIntent?: string | null;
  summary?: string | null;
  propertyInMind?: string | null;
  bedrooms?: number | null;
  budget?: string | null;
}